Unbreak minor versioning (#4)
Since minor version changes ought to be backwards compatible, that implies a publisher must be able to publish new minor version without any changes to consumers. However, that's not currently possible since validation would fail on missing schema for the new minor version. This is fixed by ensuring that only major version is part of the schema file, and any changes are applied in-place.